# public/openapi.yaml
info:
  description: ChatGPT plugin for local Gode Mode
  title: ChatGPTGodeMode
  version: v1
openapi: 3.0.1
paths:
...
  /api/saveCommand:
    post:
      description: Save last run command for later reuse, provide name and description. Name should be short command like. Description should contain when this command should be used later
      operationId: saveLastCommand
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  description: Name of the command.
                  type: string
                description:
                  description: Description of what the command does and when it should be used
                  type: string
              required:
              - name
              - description
              type: object
        required: true
      responses:
        200:
          description: Command saved successfully.
        400:
          description: Bad request.
        500:
          description: Internal server error.
servers:
- url: http://localhost:3000